Revert "refactor(columnar): align NumericColumnValues.valueCount() to long"
This reverts commit 274200313b0363afb551bdbb448bb0c1433dbdd1.

+++ libs/columnar/src/main/java/org/elasticsearch/columnar/ColumNARDocValuesConsumer.java
@@ -172,7 +172,7 @@ final class ColumNARDocValuesConsumer extends DocValuesConsumer {
             }
 
             @Override
-            public long valueCount() {
+            public int valueCount() {
                 return current.values.valueCount();
             }
